Refactor: rename `CreateEditDomainDialong`
