+++ src/modules/dashboard/components/CreateEditDomainDialog/index.ts
@@ -1,1 +1,1 @@
-export { default } from './CreateEditDomainDialog';
+export { default } from './ManageDomainsDialog';
